Moves validation to a validation package
**CHANGELOG:**

- Moves validation functions to a validation package.
