+++ go.mod
@@ -3,7 +3,11 @@ module github.com/inkatze/gocfdi
 go 1.13
 
 require (
+	github.com/konsorten/go-windows-terminal-sequences v1.0.2 // indirect
 	github.com/sirupsen/logrus v1.4.2
 	github.com/stretchr/testify v1.4.0 // indirect
 	github.com/tiaguinho/gosoap v1.2.1-0.20190910083653-a53d7a4c78d6
+	golang.org/x/net v0.0.0-20200114155413-6afb5195e5aa // indirect
+	golang.org/x/sys v0.0.0-20200113162924-86b910548bc1 // indirect
+	golang.org/x/text v0.3.2 // indirect
 )
